pyproject.toml
[project]
name = "mitre-agentic-threat-investigation"
version = "0.1.0"
description = "An AI-powered threat investigation assistant leveraging MITRE's ATT&CK framework and MCP server."
readme = "README.md"
authors = [
    { name = "Inoussa Mouiche", email = "mouiche@uwindsor.ca" }
]
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.28.1",
  "langgraph>=1.0.5",
  "pydantic>=2.12.5",
  "python-dotenv>=1.2.1",
  "rich>=14.2.0",
  "mcp[cli]>=1.1.0",
  "openai>=1.0.0",
  "mitre-mcp-server>=0.1.1",
]


[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mitre_agentic"]

[tool.uv]
package = true

